[FEAT] Constant 추가 및 cellModel 분리
