+++ oewoboka/oewoboka/Helper/Constant.swift
@@ -1,0 +1,19 @@
+//
+//  Constant.swift
+//  oewoboka
+//
+//  Created by SeoJunYoung on 2023/09/25.
+//
+
+import Foundation
+import UIKit
+
+struct Constant {
+    static let screenWidth = UIScreen.main.bounds.width
+    static let screenHeight = UIScreen.main.bounds.height
+    static let defalutPadding: CGFloat = 16
+}
+
+extension Constant {
+    
+}
